@@ -12,6 +12,7 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.network.NetworkWriter;
 import org.matsim.api.core.v01.network.Node;
+import org.matsim.application.prepare.counts.CreateCountsFromBAStData;
 import org.matsim.application.prepare.pt.CreateTransitScheduleFromGtfs;
 import org.matsim.contrib.bicycle.BicycleUtils;
 import org.matsim.contrib.osm.networkReader.LinkProperties;
@@ -43,6 +44,9 @@ import java.util.stream.Collectors;
 
 import static org.matsim.prepare.RuhrUtils.*;
 
+/**
+ * Create supply (network, pt, counts) for the scenario.
+ */
 public class CreateSupply {
 
 	public enum NetworkResolution {Low, Medium, High}
@@ -55,17 +59,19 @@ public class CreateSupply {
 
 	private static final Path nrwShape = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/shp-files/nrw/dvg2bld_nw.shp");
 
-	private static final Path gtfsData1 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/vrr_20211118_gtfs_vrr_shapes.zip");
-	private static final Path gtfsData2 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-nwl-20210215.zip");
-	private static final Path gtfsData3 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-schienenfernverkehr-de_2021-08-19.zip");
+	//new gtfs file provided by the rvr
+	private static final Path gtfsData1 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/20230106_gtfs_nrw_neue_service_ids.zip");
+	//private static final Path gtfsData1 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/vrr_20211118_gtfs_vrr_shapes.zip");
+	//private static final Path gtfsData2 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-nwl-20210215.zip");
+	private static final Path gtfsData2 = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/original-data/gtfs/gtfs-schienenfernverkehr-de_2021-08-19.zip");
 
-	private static final String gtfsDataDate1 = "2021-11-17";
-	private static final String gtfsDataDate2 = "2021-02-04";
-	private static final String gtfsDataDate3 = "2021-08-19";
+	private static final String gtfsDataDate1 = "2023-01-17";
+	//private static final String gtfsDataDate2 = "2021-02-04";
+	private static final String gtfsDataDate2 = "2021-08-19";
 
-	private static final String gtfsData1Prefix = "vrr";
-	private static final String gtfsData2Prefix = "nwl";
-	private static final String gtfsData3Prefix = "fern";
+	private static final String gtfsData1Prefix = "nrw";
+	//private static final String gtfsData2Prefix = "nwl";
+	private static final String gtfsData2Prefix = "fern";
 
 	private static final Path inputShapeNetwork1 = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-03-05_radwegeverbindungen_VM_Freizeitnetz/2021-03-05_radwegeverbindungen_VM_Freizeitnetz.shp");
 	private static final Path inputShapeNetwork2 = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-03-05_radwegeverbindungen_VM_Knotenpunktnetz/2021-03-05_radwegeverbindungen_VM_Knotenpunktnetz.shp");
@@ -73,9 +79,11 @@ public class CreateSupply {
 	// for now, we will focus on the 'Bestandsnetz'. Once, we are done with calibration, we will also generate the network for the 'Zielnetz' by replacing the previous line with the following.
 	// private static final Path inputShapeNetwork3 = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/original-data/2021-08-19_radwegeverbindungen_RRWN_Bestandsnetz_Zielnetz/2021-08-19_RRWN_Bestandsnetz_Zielnetz.shp");
 
-	private static final Path outputDirPublic = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v1.0/input/");
+	private static final Path outputDirPublic = Paths.get("public-svn/matsim/scenarios/countries/de/metropole-ruhr/metropole-ruhr-v2.0/input/");
 
-	private static final Path outputDirCounts = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v1.0/input/");
+	private static final Path outputDirCounts = Paths.get("shared-svn/projects/matsim-metropole-ruhr/metropole-ruhr-v2.0/input/");
+
+	private static final Path bastCountsRoot = Paths.get("shared-svn/projects/rvr-metropole-ruhr/data/BASt");
 	private static final Path longTermCountsRoot = Paths.get("shared-svn/projects/matsim-ruhrgebiet/original_data/counts/long_term_counts");
 	private static final Path longTermCountsIdMapping = Paths.get("shared-svn/projects/matsim-ruhrgebiet/original_data/counts/mapmatching/countId-to-nodeId-long-term-counts.csv");
 	private static final Path shortTermCountsRoot = Paths.get("shared-svn/projects/matsim-ruhrgebiet/original_data/counts/short_term_counts");
@@ -129,7 +137,11 @@ public class CreateSupply {
 
 		// ----------------------------------------- Create Network ----------------------------------------------------
 
+		// ruhr area is rather large, with a mixture of urban and more rural areas
+		// based on the paper "Road network free flow speed estimation using microscopic simulation and point-to-point travel times"
+		// the free speed factor was set to a mean of urban, metropolitan and rural areas
 		var networkBuilder = new OsmBicycleReader.Builder()
+				.setFreeSpeedFactor(0.75)
 				.setCoordinateTransformation(transformation)
 				.setIncludeLinkAtCoordWithHierarchy((coord, level) -> isIncludeLink(coord, level, ruhrGeometries, nrwGeometries))
 				.setPreserveNodeWithId(nodeIdsToKeep::contains)
@@ -159,15 +171,15 @@ public class CreateSupply {
 		// ----------------------------- Add bicycles and write network ------------------------------------------------
 
 		Network network1 = new ShpToNetwork().run(rootDirectory.resolve(inputShapeNetwork1));
-		new NetworkWriter(network1).write(outputDir.resolve("metropole-ruhr-v1.4.network-onlyBikeNetwork1.xml.gz").toString());
+		new NetworkWriter(network1).write(outputDir.resolve("metropole-ruhr-v2.0.network-onlyBikeNetwork1.xml.gz").toString());
 		new BikeNetworkMerger(network).mergeBikeHighways(network1);
 
 		Network network2 = new ShpToNetwork().run(rootDirectory.resolve(inputShapeNetwork2));
-		new NetworkWriter(network2).write(outputDir.resolve("metropole-ruhr-v1.4.network-onlyBikeNetwork2.xml.gz").toString());
+		new NetworkWriter(network2).write(outputDir.resolve("metropole-ruhr-v2.0.network-onlyBikeNetwork2.xml.gz").toString());
 		new BikeNetworkMerger(network).mergeBikeHighways(network2);
 
 		Network network3 = new ShpToNetwork().run(rootDirectory.resolve(inputShapeNetwork3));
-		new NetworkWriter(network3).write(outputDir.resolve("metropole-ruhr-v1.4.network-onlyBikeNetwork3.xml.gz").toString());
+		new NetworkWriter(network3).write(outputDir.resolve("metropole-ruhr-v2.0.network-onlyBikeNetwork3.xml.gz").toString());
 		new BikeNetworkMerger(network).mergeBikeHighways(network3);
 
 		var cleaner = new MultimodalNetworkCleaner(network);
@@ -266,19 +278,19 @@ public class CreateSupply {
 			}
 		}
 
-		String networkOut = outputDir.resolve("metropole-ruhr-v1.4.network_resolution" + networkResolution + ".xml.gz").toString();
+		String networkOut = outputDir.resolve("metropole-ruhr-v2.0.network_resolution" + networkResolution + ".xml.gz").toString();
 		new NetworkWriter(network).write(networkOut);
 
 		// --------------------------------------- Create Pt -----------------------------------------------------------
 
 		new CreateTransitScheduleFromGtfs().execute(
-				rootDirectory.resolve(gtfsData1).toString(), rootDirectory.resolve(gtfsData2).toString(), rootDirectory.resolve(gtfsData3).toString(),
-				"--date", gtfsDataDate1, gtfsDataDate2, gtfsDataDate3,
-				"--prefix", gtfsData1Prefix + "," + gtfsData2Prefix + "," + gtfsData3Prefix,
+				rootDirectory.resolve(gtfsData1).toString(), rootDirectory.resolve(gtfsData2).toString(),
+				"--date", gtfsDataDate1, gtfsDataDate2,
+				"--prefix", gtfsData1Prefix + "," + gtfsData2Prefix + "," + gtfsData2Prefix,
 				"--target-crs", "EPSG:25832",
 				"--network", networkOut,
 				"--output", outputDir.toString(),
-				"--name", "metropole-ruhr-v1.4"
+				"--name", "metropole-ruhr-v2.0"
 		);
 
 		// --------------------------------------- Create Counts -------------------------------------------------------
@@ -303,7 +315,20 @@ public class CreateSupply {
 				.build()
 				.run();
 
-		CombinedCountsWriter.writeCounts(outputDirForCounts.resolve("metropole-ruhr-v1.4.counts.xml.gz"),
+
+		new CreateCountsFromBAStData().execute(
+			"--network", networkOut,
+			"--year", "2022",
+			"--counts-mapping", rootDirectory.resolve(bastCountsRoot).resolve("bast-mapping.csv").toString(),
+			"--motorway-data",  rootDirectory.resolve(bastCountsRoot).resolve("2022_A_S.zip").toString(),
+			"--primary-data", rootDirectory.resolve(bastCountsRoot).resolve("2022_B_S.zip").toString(),
+			"--station-data", rootDirectory.resolve(bastCountsRoot).resolve("Jawe2022.csv").toString(),
+			"--output", rootDirectory.resolve(outputDirPublic).resolve("metropole-ruhr-v2.0.counts.xml.gz").toString()
+		);
+
+
+		// Based on 2015 count data, won't be used in the new model
+		CombinedCountsWriter.writeCounts(outputDirForCounts.resolve("metropole-ruhr-v2.0.counts-old.xml.gz"),
 				longTermCounts.get(RawDataVehicleTypes.Pkw.toString()), shortTermCounts.get(RawDataVehicleTypes.Pkw.toString()));
 	}
 
@@ -340,9 +365,9 @@ public class CreateSupply {
 			link.setAllowedModes(modes);
 		}
 
-		// all regular bike links should have an infrastructure speed factor of 0.5
+		// all regular bike links should have an infrastructure speed factor of 0.64 according to the rvr
 		if (link.getAllowedModes().contains(TransportMode.bike)) {
-			link.getAttributes().putAttribute(BicycleUtils.BICYCLE_INFRASTRUCTURE_SPEED_FACTOR, 0.5);
+			link.getAttributes().putAttribute(BicycleUtils.BICYCLE_INFRASTRUCTURE_SPEED_FACTOR, 0.64);
 		}
 	}
 
