@@ -4,11 +4,12 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.matsim.api.core.v01.Identifiable;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.counts.Counts;
 import org.matsim.counts.CountsWriter;
 
-public class CombinedCountsWriter<T> {
+public class CombinedCountsWriter<T extends Identifiable<T>> {
 
 	private List<Counts<T>> countsList = new ArrayList<>();
 
